requirements.txt
streamlit==1.29.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pandas==2.0.3
pyspellchecker==0.7.2
nltk==3.9.1 